Move all appium specific commands to .appium ns.
