+++ lib/api/protocol/appium/getCurrentActivity.js
@@ -1,15 +1,15 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Get the name of the current Android activity.
  *
- * @syntax .getCurrentActivity([callback])
+ * @syntax .appium.getCurrentActivity([callback])
  * @method getCurrentActivity
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {string} Name of the current activity.
- * @see getCurrentPackage
- * @see startActivity
- * @api protocol.mobile
+ * @see appium.getCurrentPackage
+ * @see appium.startActivity
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
