+++ lib/api/protocol/appium/getCurrentPackage.js
@@ -1,15 +1,15 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Get the name of the current Android package.
  *
- * @syntax .getCurrentPackage([callback])
+ * @syntax .appium.getCurrentPackage([callback])
  * @method getCurrentPackage
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {string} Name of the current package.
- * @see getCurrentActivity
- * @see startActivity
- * @api protocol.mobile
+ * @see appium.getCurrentActivity
+ * @see appium.startActivity
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
