+++ lib/api/protocol/appium/getGeolocation.js
@@ -1,16 +1,14 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Get the current geolocation of the mobile device.
- * 
- * This command works with Appium only. Getting the current geolocation for desktop browsers is not supported.
  *
- * @syntax .getGeolocation([callback])
+ * @syntax .appium.getGeolocation([callback])
  * @method getGeolocation
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {object} The current geolocation: `{latitude: number, longitude: number, altitude: number}`.
- * @see setGeolocation
- * @api protocol.mobile
+ * @see appium.setGeolocation
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
