+++ lib/api/protocol/appium/hideKeyboard.js
@@ -1,15 +1,13 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Hide soft keyboard.
  *
- * This command works with Appium only.
- *
- * @syntax .hideKeyboard([callback])
+ * @syntax .appium.hideKeyboard([callback])
  * @method hideKeyboard
  * @param {function} [callback] Optional callback function to be called when the command finishes.
- * @see isKeyboardShown
- * @api protocol.mobile
+ * @see appium.isKeyboardShown
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
