+++ lib/api/protocol/appium/isKeyboardShown.js
@@ -1,16 +1,14 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Whether or not the soft keyboard is shown.
  *
- * This command works with Appium only.
- *
- * @syntax .isKeyboardShown([callback])
+ * @syntax .appium.isKeyboardShown([callback])
  * @method isKeyboardShown
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {boolean} True if the keyboard is shown.
- * @see hideKeyboard
- * @api protocol.mobile
+ * @see appium.hideKeyboard
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
