+++ lib/api/protocol/appium/setOrientation.js
@@ -1,12 +1,14 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
- * Sets the browser orientation.
+ * Set the current device orientation.
  *
- * @param {string} orientation The new browser orientation: {LANDSCAPE|PORTRAIT}
+ * @syntax .appium.setOrientation(orientation, [callback])
+ * @method setOrientation
+ * @param {string} orientation The new device orientation: `LANDSCAPE` or `PORTRAIT`.
  * @param {function} [callback] Optional callback function to be called when the command finishes.
- * @api protocol.mobile
- * @deprecated In favour of `.appium.setOrientation()`
+ * @see appium.getOrientation
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   static get isTraceable() {
