+++ lib/api/protocol/appium/startActivity.js
@@ -1,4 +1,4 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Start an Android activity by providing package name, activity name and other optional parameters.
@@ -23,13 +23,13 @@ const ProtocolAction = require('./_base-action.js');
  *   }
  * };
  *
- * @syntax .startActivity(opts, [callback])
+ * @syntax .appium.startActivity(opts, [callback])
  * @method startActivity
  * @param {string} opts Options to start the activity with. `appPackage` and `appActivity` are required, [others](https://appium.io/docs/en/commands/device/activity/start-activity/#json-parameters) are optional.
  * @param {function} [callback] Optional callback function to be called when the command finishes.
- * @see getCurrentActivity
- * @see getCurrentPackage
- * @api protocol.mobile
+ * @see appium.getCurrentActivity
+ * @see appium.getCurrentPackage
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(opts = {}, callback) {
