+++ lib/api/protocol/getOrientation.js
@@ -1,14 +1,12 @@
 const ProtocolAction = require('./_base-action.js');
 
 /**
- * Get the current browser/device orientation.
+ * Get the current browser orientation.
  *
- * @syntax .getOrientation([callback])
- * @method getOrientation
  * @param {function} callback Callback function which is called with the result value.
- * @returns {string} The current browser/device orientation: `LANDSCAPE` or `PORTRAIT`.
- * @see setOrientation
+ * @returns {string} The current browser orientation: {LANDSCAPE|PORTRAIT}
  * @api protocol.mobile
+ * @deprecated In favour of `.appium.getOrientation()`
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
